+++ internal/models/models.go
@@ -1,0 +1,9 @@
+package models
+
+type Numbers struct {
+	Values []float64 `json:"values"`
+}
+
+type SumResponse struct {
+	Sum float64 `json:"sum"`
+}
