+++ internal/sum/sum.go
@@ -1,0 +1,9 @@
+package sum
+
+func CalculateSum(values []float64) float64 {
+	sum := 0.0
+	for _, v := range values {
+		sum += v
+	}
+	return sum
+}
